feat: :children_crossing: Write function that downloads profile data into output dir with the rest of the downloaded things.
Write a function for downloading the targeted user's profile using the download_profiles function instead of the deprecated download_profile. This is a hack around the profile data being downloaded into the root directory, and we use a temp dir with shutil to download into the temp dir  and then move the downloaded files into a 'profile' directory in the data dir with the rest of the data.